requirements.txt
asgiref==3.8.1
Django==5.2.1
django-debug-toolbar==5.2.0
Faker==37.3.0
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
